server/requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.44
botocore==1.31.44
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
contextlib2==21.6.0
cryptography==41.0.3
dill==0.3.7
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
# django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.3
fsspec==2023.6.0
google-pasta==0.2.0
gunicorn==21.2.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mpmath==1.3.0
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pandas==2.1.0
pathos==0.3.1
platformdirs==3.10.0
pox==0.3.3
ppft==1.7.6.7
protobuf==4.24.3
# psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.2
s3transfer==0.6.2
safetensors==0.3.3
sagemaker==2.184.0
schema==0.7.5
six==1.16.0
smdebug-rulesconfig==1.0.1
sqlparse==0.4.4
sympy==1.12
tblib==1.7.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.7.1
tzdata==2023.3
unicode==2.9
unicodedata2==15.0.0
Unidecode==1.3.6
urllib3==1.26.16
whitenoise==6.5.0
zipp==3.16.2
